+++ src/main/java/com/djarjo/jetson/JsonAccessType.java
@@ -1,0 +1,6 @@
+package com.djarjo.jetson;
+
+public enum JsonAccessType {
+	FIELD,
+	PROPERTY
+}