+++ src/main/java/com/djarjo/jetson/JsonDecoder.java
@@ -24,6 +24,7 @@ import java.util.logging.Level;
  * @since 2020-05-19 decode enumeration value auto-strips leading class name if present
  */
 public class JsonDecoder {
+	public final static String METHOD_FROM_JSON = "fromJson";
 	private final static FluentLogger logger = FluentLogger.forEnclosingClass();
 
 	private Set<String> _keysToSkip = null;
@@ -58,7 +59,7 @@ public class JsonDecoder {
 	 * @return a new list or a new map
 	 * @throws IllegalAccessException if a decoded value cannot be set into the target
 	 * object
-	 * @throws ParseException in case of an error in the Json string
+	 * @throws ParseException in case of an error in the JSON string
 	 */
 	public Object decode(
 			String jsonString ) throws IllegalAccessException, ParseException {
@@ -66,28 +67,47 @@ public class JsonDecoder {
 	}
 
 	/**
-	 * Decodes the given Json string and writes its values into the given target object.
+	 * Decodes the given JSON string and writes its values into the given target object.
 	 * <p>
+	 * Decoding process:
+	 * <ol>
+	 *   <li>If target is {@code null} then a new Map will be created, filled and returned</li>
+	 *   <li>If the target class is annotated with {@literal @Json} then:
+	 *     <ul><li>converter will be used if present</li>
+	 *       <li>all setters will be used (default) unless
+	 *       {@code accessType=JsonAccessType.FIELD} is set to use all public fields</li>
+	 *       <li>Members (fields and setters) annotated with {@literal @JsonTransient}
+	 *       will be skipped</li>
+	 *     </ul></li>
+	 *   <li>If target class is <b>not</b> annotated with {@literal @Json} then all members
+	 *   (fields and getters) will be used which are annotated with {@literal @Json}</li>
+	 * </ol>
 	 * Fields (getters) in target class must be annotated with
 	 * {@link com.djarjo.jetson.Json @Json}.
 	 * </p>
 	 *
-	 * @param jsonString Json string to decode
-	 * @param target target object for the values from the Json string
+	 * @param json JSON text to decode
+	 * @param target target object to write values into
 	 * @return given target or new instance of list or map
-	 * @throws IllegalAccessException if a decoded value cannot be set into the target
-	 * object
-	 * @throws ParseException in case of an error in the Json string
+	 * @throws IllegalAccessException if a decoded value cannot be set
+	 * @throws ParseException in case of an error in the JSON string
 	 */
-	public Object decodeIntoObject( String jsonString,
+	public Object decodeIntoObject( String json,
 			Object target ) throws ParseException, IllegalAccessException {
-		if ( jsonString == null || jsonString.isEmpty() ) {
+		if ( json==null || json.isBlank() ) {
 			return null;
 		}
-		_tokenizer = new Tokenizer( jsonString,
+		_tokenizer = new Tokenizer( json,
 				TokenizerOption.SKIP_DOUBLESLASH_UNTIL_EOL,
 				TokenizerOption.SKIP_LINEBREAK, TokenizerOption.SKIP_WHITESPACE );
 		_tokenizer.nextToken();
+		Json classAnno = target.getClass().getAnnotation( Json.class );
+		if ( classAnno!=null ) {
+			Object object = _handleClassAnnotation( target, classAnno );
+			if ( object!=null ) {
+				return object;
+			}
+		}
 		return _decodeValue( target, null, null );
 	}
 
@@ -104,14 +124,14 @@ public class JsonDecoder {
 	@SuppressWarnings("unchecked")
 	public <T> List<T> decodeList( String jsonString,
 			Class<T> itemClass ) throws ParseException, IllegalAccessException {
-		if ( jsonString == null || jsonString.isEmpty() ) {
+		if ( jsonString==null || jsonString.isEmpty() ) {
 			return null;
 		}
 		_tokenizer = new Tokenizer( jsonString,
 				TokenizerOption.SKIP_DOUBLESLASH_UNTIL_EOL,
 				TokenizerOption.SKIP_LINEBREAK, TokenizerOption.SKIP_WHITESPACE );
 		Token token = _tokenizer.nextToken();
-		if ( token.symbol != Symbol.LEFT_BRACKET ) {
+		if ( token.symbol!=Symbol.LEFT_BRACKET ) {
 			String s = _makeErrorInfo( "Syntax error. '[' expected to parse a list" );
 			throw new ParseException( s, 0 );
 		}
@@ -127,7 +147,7 @@ public class JsonDecoder {
 	 * @return this for fluent API
 	 */
 	public JsonDecoder keysToSkip( String... keys ) {
-		_keysToSkip = (_keysToSkip == null) ? new HashSet<>() : _keysToSkip;
+		_keysToSkip = (_keysToSkip==null) ? new HashSet<>() : _keysToSkip;
 		Collections.addAll( _keysToSkip, keys );
 		return this;
 	}
@@ -181,27 +201,27 @@ public class JsonDecoder {
 	private Collection _decodeList( Collection target,
 			Type valueType ) throws IllegalAccessException, ParseException {
 		Token token;
-		target = (target == null) ? new ArrayList<>() : target;
+		target = (target==null) ? new ArrayList<>() : target;
 
 		//--- Build a new list
 		List<Object> listFromJson = new ArrayList<>();
 		while ( true ) {
 			token = _tokenizer.nextToken();
-			if ( token.symbol == Symbol.COMMA ) {
+			if ( token.symbol==Symbol.COMMA ) {
 				token = _tokenizer.nextToken();
 			}
-			if ( token.symbol == Symbol.RIGHT_BRACKET ) {
+			if ( token.symbol==Symbol.RIGHT_BRACKET ) {
 				break; // --- Safely skip , in front of }
 			}
-			if ( token.symbol == Symbol.EOF ) {
+			if ( token.symbol==Symbol.EOF ) {
 				String s = _makeErrorInfo( "Syntax error. Value or ']' expected" );
 				throw new ParseException( s, _tokenizer.getPosition() );
 			}
 			//--- extract json list element (if map) for optional second parse
-			int startIndex = (token.symbol == Symbol.LEFT_BRACE) ? token.position : -1;
+			int startIndex = (token.symbol==Symbol.LEFT_BRACE) ? token.position : -1;
 			//--- Value from Json string
 			Object value = _decodeValue( null, null, valueType );
-			if ( valueType != null ) {
+			if ( valueType!=null ) {
 				value = BaseConverter.convertToType( value, (Class<?>) valueType );
 			}
 			boolean isMerged = false;
@@ -251,15 +271,15 @@ public class JsonDecoder {
 			Type keyType, Type valueType ) throws ParseException,
 			IllegalAccessException {
 		Object key = null, value = null;
-		if ( target == null ) {
+		if ( target==null ) {
 			target = new HashMap<>();
 		}
 		while ( true ) {
 			key = _obtainNameFromJson();
-			if ( key == null ) {
+			if ( key==null ) {
 				break;
 			}
-			if ( keyType != null ) {
+			if ( keyType!=null ) {
 				key = BaseConverter.convertToType( key, (Class<?>) keyType );
 			}
 			value = _decodeValue( null, null, valueType );
@@ -269,48 +289,64 @@ public class JsonDecoder {
 	}
 
 	/**
-	 * Decodes a Json string which has started with "{". Decoding ends when finding the
+	 * Decodes a JSON string which has started with "{". Decoding ends when finding the
 	 * matching "}". Entries are decoded in between. Each entry looks like
 	 *
 	 * @param target The target object for the values
 	 * @param keyType type of key for a map
-	 * @param valueType generic type of a map or collection
-	 * @return target or new map filled from Json string
+	 * @param valueType generic type of map or collection
+	 * @return target or new map filled from JSON string
 	 * @throws IllegalAccessException if access to an object field or setter fails
 	 * @throws ParseException if json string has a format error
 	 */
 	@SuppressWarnings("unchecked")
 	private Object _decodeObject( Object target, Type keyType,
 			Type valueType ) throws IllegalAccessException, ParseException {
-		Map<String, Member> members = null;
-		String name = null;
-		Token token = null;
-		Object value = null;
+		boolean accessFields = false;
+		boolean allMembers = false;
+		Map<String, Member> members;
+		String name;
+		Token token;
+		Object value;
 
 		// --- New target if none given
-		if ( target == null ) {
-			if ( valueType == null || valueType == Object.class ) {
+		if ( target==null ) {
+			if ( valueType==null || valueType==Object.class ) {
 				return _decodeMap( _tokenizer, null, null, valueType );
 			}
 			target = ReflectionHelper.createInstanceFromType( valueType );
 		}
 
-		// --- target is a map or a class with fields and getters
+		// --- target is a map
 		if ( Map.class.isAssignableFrom( target.getClass() ) ) {
 			return _decodeMap( _tokenizer, (Map<Object, Object>) target, keyType,
 					valueType );
 		}
+
+		//--- Check if class is annotated with @Json
+		Json classAnno = target.getClass().getAnnotation( Json.class );
+		if ( classAnno!=null ) {
+			Object object = _handleClassAnnotation( target, classAnno );
+			if ( object!=null ) {
+				return object;
+			}
+			if ( JsonAccessType.FIELD.equals( classAnno.accessType() ) ) {
+				accessFields = true;
+			}
+			allMembers = true;
+		}
+
 		members = _obtainMembers( target );
 
-		// --- Loop through Json object until "}"
+		// --- Loop through JSON object until "}"
 		while ( true ) {
 			name = _obtainNameFromJson();
-			if ( name == null ) {
+			if ( name==null ) {
 				break;
 			}
 			Member member = members.get( name );
-			if ( member == null ) {
-				if ( (_logLevel4MissingAttribute != Level.OFF) && (!_keysToSkip.contains(
+			if ( member==null ) {
+				if ( (_logLevel4MissingAttribute!=Level.OFF) && (!_keysToSkip.contains(
 						name )) ) {
 					String message =
 							_makeErrorInfo(
@@ -324,7 +360,7 @@ public class JsonDecoder {
 			Json anno = member.getClass().getAnnotation( Json.class );
 			// --- Next token must be the value. Delay target instantiation
 			token = _tokenizer.getToken();
-			if ( token.symbol == Symbol.LEFT_BRACKET || token.symbol == Symbol.LEFT_BRACE ) {
+			if ( token.symbol==Symbol.LEFT_BRACKET || token.symbol==Symbol.LEFT_BRACE ) {
 				value = _decodeObjectValue( target, member );
 			} else {
 				value = _decodeValue( null, null, null );
@@ -357,19 +393,19 @@ public class JsonDecoder {
 			if ( Collection.class.getName()
 					.equals( rawTypeName ) || List.class.getName()
 					.equals( rawTypeName ) ) {
-				if ( newTarget == null ) {
+				if ( newTarget==null ) {
 					newTarget = new ArrayList<>();
 				}
 				valueType = ReflectionHelper.getActualTypeArgument( pType, 0 );
 			} else if ( Set.class.getName()
 					.equals( rawTypeName ) ) {
-				if ( newTarget == null ) {
+				if ( newTarget==null ) {
 					newTarget = new HashSet<>();
 				}
 				valueType = ReflectionHelper.getActualTypeArgument( pType, 0 );
 			} else if ( Map.class.getName()
 					.equals( rawTypeName ) ) {
-				if ( newTarget == null ) {
+				if ( newTarget==null ) {
 					newTarget = new HashMap<>();
 				}
 				keyType = ReflectionHelper.getActualTypeArgument( pType, 0 );
@@ -378,12 +414,12 @@ public class JsonDecoder {
 					.contentEquals( rawTypeName ) ) {
 				keyType = ReflectionHelper.getActualTypeArgument( pType, 0 );
 				valueType = ReflectionHelper.getActualTypeArgument( pType, 1 );
-				if ( newTarget == null ) {
+				if ( newTarget==null ) {
 					newTarget = new EnumMap( (Class<?>) keyType );
 				}
 			}
 		}
-		if ( newTarget == null ) {
+		if ( newTarget==null ) {
 			valueType = type;
 		}
 		return _decodeValue( newTarget, keyType, valueType );
@@ -406,20 +442,20 @@ public class JsonDecoder {
 	 * @param target target object
 	 * @param keyType type of key in a map ({@code null} if target is no map)
 	 * @param valueType type of value
-	 * @return value from decoded Json string
+	 * @return value from decoded JSON string
 	 * @throws IllegalAccessException on target access error
 	 * @throws ParseException on json format error
 	 */
 	private Object _decodeValue( Object target, Type keyType,
 			Type valueType ) throws IllegalAccessException, ParseException {
-		Object retval = null;
+		Object retval;
 		Token token = _tokenizer.getToken();
 
-		if ( token.symbol == Symbol.LEFT_BRACKET ) { // Json List
+		if ( token.symbol==Symbol.LEFT_BRACKET ) { // JSON List
 			retval = _decodeList( (Collection<?>) target, valueType );
-		} else if ( token.symbol == Symbol.LEFT_BRACE ) { // --- Json Object
+		} else if ( token.symbol==Symbol.LEFT_BRACE ) { // JSON Object
 			retval = _decodeObject( target, keyType, valueType );
-		} else if ( token.symbol == Symbol.WORD ) {
+		} else if ( token.symbol==Symbol.WORD ) {
 			String name = token.value;
 			if ( name.equalsIgnoreCase( "false" ) ) {
 				retval = Boolean.FALSE;
@@ -434,33 +470,85 @@ public class JsonDecoder {
 		}
 
 		// --- Value is an integer
-		else if ( token.symbol == Symbol.VALUE_INTEGER ) {
+		else if ( token.symbol==Symbol.VALUE_INTEGER ) {
 			retval = token.getAsLong();
 		}
 
 		// --- Value is hexadecimal
-		else if ( token.symbol == Symbol.VALUE_HEX ) {
+		else if ( token.symbol==Symbol.VALUE_HEX ) {
 			retval = TextHelper.parseHex( token.value );
 		}
 
 		// --- Value is a decimal or double
-		else if ( token.symbol == Symbol.VALUE_DECIMAL ) {
+		else if ( token.symbol==Symbol.VALUE_DECIMAL ) {
 			retval = Double.parseDouble( token.value );
 		}
 
 		// --- Value is a decimal
-		else if ( token.symbol == Symbol.VALUE_DOUBLE ) {
+		else if ( token.symbol==Symbol.VALUE_DOUBLE ) {
 			retval = token.getAsDouble();
 		}
 
 		// --- Value is a string
-		else if ( token.symbol == Symbol.STRING ) {
+		else if ( token.symbol==Symbol.STRING ) {
 			retval = token.value;
 		} else {
 			String s = _makeErrorInfo( "Unknown token " + token );
 			throw new ParseException( s, _tokenizer.getPosition() );
 		}
 		return retval;
+	}
+
+	/**
+	 * Cuts out JSON object from tokenizer current position until matching character
+	 *
+	 * @return substring of JSON text
+	 */
+	private String _cutOutUntilMatching() {
+		Token token = _tokenizer.getToken();
+		if ( token.symbol==Symbol.STRING ) {
+			return token.value;
+		}
+		token = _tokenizer.clipUntilMatching();
+		return token.value;
+	}
+
+	/**
+	 * Handles a target class which is annotated with {@literal @Json}.
+	 * <p>
+	 * If {@literal @Json} specifies a converter then the current section from the tokenizer
+	 * will be cut out and given to converters decode method.
+	 * </p><p>
+	 * If the class contains a method named {@code fromJson(String)} then the current
+	 * section from the tokenizer will be cut out and given to that method.
+	 * </p>
+	 *
+	 * @param target current bean
+	 * @param anno class level annotation {@literal @Json}
+	 * @return return value from converter or {@code fromJson} or {@code null} if none exists
+	 */
+	@SuppressWarnings("rawtypes")
+	private Object _handleClassAnnotation( Object target, Json anno ) {
+		// --- use converter if given
+		JsonConverter converter = Jetson.getConverter( anno );
+		if ( converter!=null ) {
+			String jsonText = _cutOutUntilMatching();
+			return converter.decodeFromJson( jsonText );
+		}
+
+		//--- use method "fromJson()" if it exists in POJO class
+		Method fromJson = ReflectionHelper.findMethod( target.getClass(), METHOD_FROM_JSON );
+		if ( fromJson!=null ) {
+			String jsonText = _cutOutUntilMatching();
+			try {
+				return fromJson.invoke( target, jsonText );
+			} catch ( IllegalAccessException | InvocationTargetException e ) {
+				logger.atWarning().withCause( e )
+						.log( "Cannot invoke %s.fromJson( %s )", target.getClass().getName(),
+								jsonText );
+			}
+		}
+		return null;
 	}
 
 	/**
@@ -492,7 +580,7 @@ public class JsonDecoder {
 		for ( Field field : target.getClass()
 				.getFields() ) {
 			Json anno = field.getAnnotation( Json.class );
-			if ( anno != null ) {
+			if ( anno!=null ) {
 				String name = anno.key();
 				if ( name.equals( Json.defaultName ) ) {
 					name = field.getName();
@@ -505,7 +593,7 @@ public class JsonDecoder {
 		for ( Method method : target.getClass()
 				.getMethods() ) {
 			Json anno = method.getAnnotation( Json.class );
-			if ( anno != null ) {
+			if ( anno!=null ) {
 				String name = anno.key();
 				if ( name.equals( Json.defaultName ) ) {
 					name = ReflectionHelper.getVarNameFromMethodName( method.getName() );
@@ -527,20 +615,20 @@ public class JsonDecoder {
 		Token token = _tokenizer.nextToken();
 
 		// --- '}' ends parsing
-		if ( token.symbol == Symbol.RIGHT_BRACE ) {
+		if ( token.symbol==Symbol.RIGHT_BRACE ) {
 			return null;
 		}
 
 		// --- Skip comma ',' as separator between elements
-		if ( token.symbol == Symbol.COMMA ) {
+		if ( token.symbol==Symbol.COMMA ) {
 			token = _tokenizer.nextToken();
 		}
-		if ( token.symbol == Symbol.RIGHT_BRACE ) {
+		if ( token.symbol==Symbol.RIGHT_BRACE ) {
 			return null; // --- Safely skip , in front of }
 		}
 
 		// --- "name" expected
-		if ( (token.symbol != Symbol.STRING) && (token.symbol != Symbol.WORD) ) {
+		if ( (token.symbol!=Symbol.STRING) && (token.symbol!=Symbol.WORD) ) {
 			throw new JsonParseException(
 					_makeErrorInfo( "Syntax error. Name or '}' expected" ) );
 		}
@@ -548,7 +636,7 @@ public class JsonDecoder {
 
 		// --- ':' expected
 		token = _tokenizer.nextToken();
-		if ( token.symbol != Symbol.COLON ) {
+		if ( token.symbol!=Symbol.COLON ) {
 			throw new JsonParseException(
 					_makeErrorInfo( "Syntax error. Colon ':' expected" ) );
 		}
@@ -578,7 +666,7 @@ public class JsonDecoder {
 
 		if ( anno.decodable() ) {
 			// --- Use "converter" if present
-			if ( anno.converter() != JsonConverter.class ) {
+			if ( anno.converter()!=JsonConverter.class ) {
 				value = _useConverterToDecode( anno, value );
 			}
 
@@ -606,9 +694,9 @@ public class JsonDecoder {
 	 */
 	private void _skipValue() {
 		Token token = _tokenizer.getToken();
-		if ( token.symbol == Symbol.LEFT_BRACKET ) {
+		if ( token.symbol==Symbol.LEFT_BRACKET ) {
 			_skipCollection();
-		} else if ( token.symbol == Symbol.LEFT_BRACE ) {
+		} else if ( token.symbol==Symbol.LEFT_BRACE ) {
 			_skipMap();
 		}
 	}
@@ -616,7 +704,7 @@ public class JsonDecoder {
 	@SuppressWarnings("rawtypes")
 	private Object _useConverterToDecode( Json anno,
 			Object value ) throws IllegalAccessException {
-		if ( (anno != null) && (value != null) && (anno.converter() != JsonConverter.class) ) {
+		if ( (anno!=null) && (value!=null) && (anno.converter()!=JsonConverter.class) ) {
 			Class<? extends JsonConverter> annotatedConverter = anno.converter();
 			try {
 				JsonConverter converter = annotatedConverter.getDeclaredConstructor()