+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0
+
+* Classes can now be annotated with @Json to encode all members. Class level annotation uses accessType to distinguish
+  between fields and properties
+* New annotation @JsonTransient to be used on fields and properties to suppress decoding and encoding
+* Added method `decodeToList( String, Class)` to support decoding lists outside of pojos (which have no generic)
+
 ## 1.1.0
 
 * Decoding a collection into an object overwrites current collection